Improve the deduplication script
With this PR the deduplication script is improved, spelling mistakes are corrected in identify_duplicates and merge_datasets. Moreover a script was created to call for these functions and apply them on the datasets. The latter is also a work in progress. When it is done it will be the script that needs to be run to get the final dataset.