Add corrected_inclusion columns
With this commit the quality_check function is expanded so that it creates 3 new columns, which will contain the corrected labels after two quality checks (1) excluded records that should have been included and (2) vice versa.
- anxiety_included_corrected
- depression_included_corrected
- substance_included_corrected

Up until now they are only filled according to the first quality check, but hopefully later today I will also manage to do so for the second quality check.

problem to solve for quality check 2:
It is a bit hard to match the rows from the depression quality check 2 dataset, because there are many empty cells. Therefore I will work on a solution where the matching is based first on title (which is also done for the other subjects) and then for the missing titles, on record_id (that's the only known variable). However, the problem with record_id is that sometimes the record_id is not unique in the large merged dataset. Record_id's were created per subject, so after merging the same record_id may refer to two or more different records.

Moreover, I adapted the deduplicate_titles.R script to be able to handle missing titles.