Finalize quality_check.R script
Yay, it works! As rows should not have only contained the record id's the script was already basically finished within the previous commit. The former part of quality check 1 could be recycled for quality check 2.

Within the deduplicate_titles.R a fail-safe is added, to make sure that the script is only ran when there are indeed duplicates. Without this extra statement the quality_check function breaks when a there are no duplicate titles found when matching the merged dataset with the quality_check set.

In master_script_merging_after_asreview.R The information on what happens in the quality check is added as well as adding some finishing touches to the resulting dataset.