Create quality check function and title deduplication (#13)
* Start on quality check function and title deduplication

This commit makes the following changes:
- composite_label is now more explicit
- Deduplication is now also applied when a record has a composite label of 0, there is either a 0 or 1 present in one of the subjects, and the already existing statements
- A start on code for the quality checks (correcting for incorrect labels)
- Deduplication based on title (some of the quality check records do not have doi, hence matching will be done via titles)
- Update the master_script.

* Add corrected_inclusion columns

With this commit the quality_check function is expanded so that it creates 3 new columns, which will contain the corrected labels after two quality checks (1) excluded records that should have been included and (2) vice versa.
- anxiety_included_corrected
- depression_included_corrected
- substance_included_corrected

Up until now they are only filled according to the first quality check, but hopefully later today I will also manage to do so for the second quality check.

problem to solve for quality check 2:
It is a bit hard to match the rows from the depression quality check 2 dataset, because there are many empty cells. Therefore I will work on a solution where the matching is based first on title (which is also done for the other subjects) and then for the missing titles, on record_id (that's the only known variable). However, the problem with record_id is that sometimes the record_id is not unique in the large merged dataset. Record_id's were created per subject, so after merging the same record_id may refer to two or more different records.

Moreover, I adapted the deduplicate_titles.R script to be able to handle missing titles.

* Finalize quality_check.R script

Yay, it works! As rows should not have only contained the record id's the script was already basically finished within the previous commit. The former part of quality check 1 could be recycled for quality check 2.

Within the deduplicate_titles.R a fail-safe is added, to make sure that the script is only ran when there are indeed duplicates. Without this extra statement the quality_check function breaks when a there are no duplicate titles found when matching the merged dataset with the quality_check set.

In master_script_merging_after_asreview.R The information on what happens in the quality check is added as well as adding some finishing touches to the resulting dataset.